Initial working version of spine-in-gui implementation.

+++ engine/rig/src/rig.h
@@ -228,8 +228,8 @@ namespace dmRig
         float x;
         float y;
         float z;
-        uint16_t u;
-        uint16_t v;
+        float u;
+        float v;
         union  {
             struct {
                 uint8_t r;
